Move tests into src/snark/varuna/sage_test.rs
