+++ algorithms/src/snark/varuna/ahp/prover/round_functions/first.rs
@@ -161,47 +161,3 @@ impl<F: PrimeField, SM: SNARKMode> AHPForR1CS<F, SM> {
 }
 
 pub type Witness<F> = LabeledPolynomial<F>;
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{
-        snark::marlin::{MarlinHidingMode, MarlinSNARK, TestCircuit},
-        traits::{algebraic_sponge::AlgebraicSponge, snark::SNARK},
-    };
-    use snarkvm_curves::bls12_377::{Bls12_377, Fq, Fr};
-    use snarkvm_fields::{One, Zero};
-    use snarkvm_utilities::Uniform;
-    use std::{borrow::Cow, collections::HashMap, ops::Deref};
-    type FS = crate::crypto_hash::PoseidonSponge<Fq, 2, 1>;
-    type MarlinSonicInst = MarlinSNARK<Bls12_377, FS, MarlinHidingMode>;
-
-    // TODO: make macros to use different hiding modes
-    #[test]
-    fn test_prover_first_round() {
-        let random_state = vec![3u64; 100];
-        let rng = &mut snarkvm_utilities::rand::TestMockRng::fixed(random_state);
-
-        let a = Fq::rand(rng);
-        println!("a: {:?}", a);
-        let a = Fq::rand(rng);
-        println!("a: {:?}", a);
-
-        let max_degree = AHPForR1CS::<Fr, MarlinHidingMode>::max_degree(100, 25, 300).unwrap();
-        let universal_srs = MarlinSonicInst::universal_setup(max_degree).unwrap();
-        let universal_prover = &universal_srs.to_universal_prover().unwrap();
-        let universal_verifier = &universal_srs.to_universal_verifier().unwrap();
-        let fs_parameters = FS::sample_parameters();
-        let mul_depth = 1;
-        let num_constraints = 100;
-        let num_variables = 25;
-        let (circ, public_inputs) = TestCircuit::gen_rand(mul_depth, num_constraints, num_variables, rng);
-        let (index_pk, index_vk) = MarlinSonicInst::circuit_setup(&universal_srs, &circ).unwrap();
-        let mut keys_to_constraints = BTreeMap::new();
-        keys_to_constraints.insert(index_pk.circuit.deref(), std::slice::from_ref(&circ));
-
-        let prover_state = AHPForR1CS::<_, MarlinHidingMode>::init_prover(&keys_to_constraints, rng).unwrap();
-        let state = AHPForR1CS::<_, MarlinHidingMode>::prover_first_round(prover_state, rng).unwrap();
-        // TODO: checks on state.
-    }
-}
