+++ algorithms/src/snark/varuna/resources/challenges.input
