+++ algorithms/src/snark/varuna/resources/inputs.txt
