+++ algorithms/src/snark/varuna/resources/test_rounds/h_0.test_vector
@@ -1,0 +1,1 @@
+"7388904030749824121217721821433853214953911918259805849443329273927733084164,2135176220032946734943829824459106118384054561849250372330716048683975588842,4754990573109988415472033418479071784044875994712276662893248150540227089791,2758334532884618923758430413601802477377254380601541511107093225976782776983,8095203738994452079955122826479915970380729551727226719303332837983196321727,6177424990788360784192413675740409023496996883035955845846493361503576983134,5094286905430846656542049153180259645877255793223574061499131126611824661186,"
